requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=1.0.0=pypi_0
blas=1.0=mkl
blessed=1.20.0=pypi_0
bottleneck=1.3.4=py37hce1f21e_0
ca-certificates=2023.05.30=h06a4308_0
cachetools=4.2.4=pypi_0
certifi=2022.12.7=py37h06a4308_0
charset-normalizer=2.0.12=pypi_0
click=8.1.3=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.11.0=pypi_0
fonttools=4.38.0=pypi_0
google-auth=1.35.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
gpustat=1.1=pypi_0
grpcio=1.46.1=pypi_0
idna=3.3=pypi_0
importlib-metadata=4.11.3=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.1.0=pypi_0
kiwisolver=1.4.5=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libstdcxx-ng=9.3.0=hd4cf53a_17
markdown=3.3.7=pypi_0
matplotlib=3.5.3=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
ncurses=6.3=h7f8727e_2
networkx=2.6.3=pyhd3eb1b0_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
nltk=3.6.2=pypi_0
numexpr=2.8.1=py37h6abb31d_0
numpy=1.20.3=py37hf144106_0
numpy-base=1.20.3=py37h74d4b33_0
nvidia-ml-py=12.535.77=pypi_0
oauthlib=3.2.0=pypi_0
openssl=1.1.1u=h7f8727e_0
packaging=22.0=py37h06a4308_0
pandas=1.3.5=py37h8c16a72_0
pillow=9.5.0=pypi_0
pip=21.2.2=py37h06a4308_0
protobuf=3.20.1=pypi_0
psutil=5.9.5=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyparsing=3.1.1=pypi_0
python=3.7.11=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.5.1=py3.7_cuda10.2.89_cudnn7.6.5_0
pytz=2022.7=py37h06a4308_0
pyyaml=6.0=pypi_0
readline=8.1.2=h7f8727e_1
regex=2022.4.24=pypi_0
requests=2.27.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=61.2.0=py37h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sklearn=0.0=pypi_0
sqlite=3.38.3=hc218d9a_0
tensorboard=2.6.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.11=h1ccaba5_1
tqdm=4.62.2=pypi_0
typing-extensions=4.2.0=pypi_0
urllib3=1.26.9=pypi_0
wcwidth=0.2.6=pypi_0
werkzeug=2.1.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7f8727e_1
zipp=3.8.0=pypi_0
zlib=1.2.12=h7f8727e_2
